feat: quantized boson charge — pion/lepton charge tunable via BOSON_CHARGE
Pion and lepton charges are now f32 with magnitude BOSON_CHARGE (default
0.1) instead of hardcoded ±1 integers. Particle charge is quantized to
nearest BOSON_CHARGE multiple at all entry points (addParticle, Roche
transfer, disintegration). Charge conservation maintained across
emission, absorption, decay, and annihilation on both CPU and GPU.

+++ src/gpu/shaders/shared-structs.wgsl
@@ -77,7 +77,7 @@ struct Photon {
 struct Pion {
     posX: f32, posY: f32,
     wX: f32, wY: f32,
-    mass: f32, charge: i32, energy: f32,
+    mass: f32, charge: f32, energy: f32,
     emitterId: u32, age: u32, flags: u32,
     kind: u32, _pad1: u32,   // kind: 0=pion, 1=lepton
 };
